Backend/requirements.txt
flask-mysqldb==0.2.0
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
mysql-connector==2.1.6